Undo bug fix
Initial problem: Player kept momentum when terminal was fixed as the terminal handled keyUp and keyDown events.

The fix was to allow keyUp events as this would allow the player to stop moving. However, this introduced a new bug where the player would be able to move in the opposite direction as momentum was never added by new keyDown events but could still be subtracted from new keyUp events.