Closes #20 Keep enemies within a set area
Enemies can now be made to wonder within a set area, to return to a set area when they leave it, and to only attack enemies within a set area. These features are designed to work together but can work well seperately.

For now the anchor is represented by a fallen crown, but it doesn't have to be visible or interactable to work.