Projectile movement and Enemy movement
Projectiles now move to a set point and ranged ghost will stay a set distance away from the player when approaching.
Arrows despawn once they hit their target location or the target.
Arrows have been integrated as weapons as they will share attributes with other future weapons

Need help fixing the location the ghosts go to when approaching the player. As of now they are restricted to the x or y axis around the player